feat: create layeredPanel
