+++ src/main/scala/ulisse/infrastructures/view/dashboard/Dashboard.scala
@@ -1,7 +1,7 @@
 package ulisse.infrastructures.view.dashboard
 import ulisse.infrastructures.view.common.Themes.Theme
 import ulisse.infrastructures.view.components.ComponentMixins.UpdatableContainer
-import ulisse.infrastructures.view.components.LayeredContainers.*
+import ulisse.infrastructures.view.components.container.LayeredContainers.*
 
 import java.awt.BorderLayout
 import scala.swing.{BorderPanel, Component}
