+++ src/main/scala/ulisse/infrastructures/view/map/MapItem.scala
@@ -1,7 +1,7 @@
 package ulisse.infrastructures.view.map
 
 import ulisse.infrastructures.view.common.Themes.Theme
-import ulisse.infrastructures.view.components.JImages.*
+import ulisse.infrastructures.view.components.draw.JImages.*
 import ulisse.infrastructures.view.components.ui.decorators.Styles
 import ulisse.infrastructures.view.map.ViewObservers.ViewObservable
 import ulisse.utils.Pair
